+++ 0x12-singly_linked_lists/lists.h
@@ -18,6 +18,6 @@ typedef struct list_s
 	struct list_s *next;
 } list_t;
 
-size_t print_list(const list_t *h);
+int print_list(const list_t *h);
 
 #endif
